requirements.txt
flask
Pillow
opencv-python
werkzeug
